requirements.txt
# --------- pytorch --------- #
torch>=1.8.0
torchvision>=0.11.0
pytorch-lightning>=1.5.10
torchmetrics>=0.7.0

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# tensorboard

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
pudb            # debugger

pillow
scipy
matplotlib
opencv-python
pycocotools

imageio
pypng
Cython
PyOpenGL
triangle
glumpy
pytz
open3d
einops